apps/python-service/src/config.py
"""Environment-backed settings."""

from __future__ import annotations

import os
from functools import lru_cache

from dotenv import load_dotenv

load_dotenv()


@lru_cache
def get_audd_api_token() -> str:
    return os.environ.get("AUDD_API_TOKEN", "").strip()


def get_cors_origins() -> list[str]:
    raw = os.environ.get("CORS_ORIGINS", "http://localhost:3000")
    return [o.strip() for o in raw.split(",") if o.strip()]
